+++ packages/envsync-enterprise/src/routes/oidc.route.ts
@@ -3,16 +3,16 @@ import { describeRoute } from "hono-openapi";
 import { resolver, validator as zValidator } from "hono-openapi/zod";
 
 import { OidcController } from "../controllers/oidc.controller";
-import { authMiddleware } from "@/middlewares/auth.middleware";
-import { enterpriseGuard } from "@/middlewares/enterprise.middleware";
-import { requirePermission } from "@/middlewares/permission.middleware";
+import { authMiddleware } from "envsync-api/ports/middlewares";
+import { enterpriseGuard } from "envsync-api/ports/middlewares";
+import { requirePermission } from "envsync-api/ports/middlewares";
 import {
 	createOidcProviderRequestSchema,
 	oidcProviderResponseSchema,
 	oidcProvidersResponseSchema,
 	updateOidcProviderRequestSchema,
 } from "../validators/oidc.validator";
-import { errorResponseSchema } from "@/validators/common";
+import { errorResponseSchema } from "envsync-api/ports/validators-common";
 
 const app = new Hono();